[conv2d/tflite] Fix uint overflow in grouped conv offsets; reject groups>1 in TFLite export
- conv2d_layer.cpp: cast first operand of group-offset multiplications to
  size_t to prevent unsigned int overflow when channels/spatial dims are large
  (g * icg * ihw, g * ocg * icg * fh * fw, g * ocg * owoh)
- node_exporter.cpp: TFLite CONV_2D has no groups field; throw explicitly
  when groups > 1 rather than silently emitting an invalid/wrong node

+++ nntrainer/layers/conv2d_layer.cpp
@@ -493,10 +493,11 @@ void Conv2DLayer::forwarding(RunLayerContext &context, bool training) {
       Tensor result = Tensor(calcCol2ImOutputDim(out_dim, fdim_g));
       for (unsigned int g = 0; g < groups; ++g) {
         Tensor in_g = in_sub.getSharedDataTensor(
-          {1, icg, in_dim.height(), in_dim.width()}, g * icg * ihw);
+          {1, icg, in_dim.height(), in_dim.width()}, (size_t)g * icg * ihw);
         Tensor filt_g = filter_kernel.getSharedDataTensor(
-          {ocg, icg * fh * fw}, g * ocg * icg * fh * fw);
-        Tensor out_g = out.getSharedDataTensor({ocg, owoh}, g * ocg * owoh);
+          {ocg, icg * fh * fw}, (size_t)g * ocg * icg * fh * fw);
+        Tensor out_g =
+          out.getSharedDataTensor({ocg, owoh}, (size_t)g * ocg * owoh);
         result.setZero();
         im2col(in_g, fdim_g, padding, stride, dilation, result);
         filt_g.dot(result, out_g, false, true);
